backend/requirements.txt
flask
flask-socketio
flask-cors
python-chess
torch
numpy
python-socketio